# Fix memory allocations on CPU