@glwagner 2020-03-06
@views allocates because it creates a new view every time it runs. Not sure if that’s the issue. To avoid this we have to preallocate the view, somehow.